backend/requirements.txt
fastapi>=0.95.0
pydantic>=2.0.0
uvicorn>=0.22.0
websockets>=11.0.0
pytest>=7.3.1
python-dotenv>=0.21.0